Docs: Fix several board-specific values.
Ugh.

Fix BOARD_NR_ADC_PINS and boardADCPins on Maple, Maple Mini, and Maple
RET6 Edition:

Maple and Maple RET6 Ed. don't have ADC on pin 3.  Seems likely that
the error was due to AIN being missing on the silkscreen.  They also
have ADC on pin 13.  This is not really true due to the voltage drop
across the built-in LED.

Maple Mini really should not include BOARD_LED_PIN among its ADC pins,
since it's not broken out to a header.

Fix BOARD_NR_PWM_PINS and boardPWMPins on Maple and RET6 Ed:

Both boards falsely include pin 25 as a PWM pin.  Maple RET6 Edition
also lacks some PWM pins.

Fix BOARD_NR_SPI on Maple RET6 Edition:

SPI3 uses PB4, which is unusable since it's tied to NRST.  This means
that SPI3 is not available on the RET6 edition.  Decreasing
BOARD_NR_SPI to 2 to accomodate this.  Leave the BOARD_SPI3_*_PIN
defines intact so as not to further clutter the rest of the library
with RET6-specific wrinkles.