openocd: Fix repeated JTAG flash failure, use alternate reset config
1) Reset, halt, and unprotect the the flash before writing to it.
This fixes a bug in which every other flash attempt would fail.

2) Maple R5 and below have NRST and JTNRST erroneously tied together,
resulting in a full TAP and system reset when a reset is asserted.
This prevents the 'reset halt' command from working.

This can be fixed hard-hack style by cutting the trace out of JTNRST.

Users of the Leaf Maple JTAG adapter will also need to cut the trace
between TRST and SRST.

3) Assuming you have a functioning 'reset halt' setup (see 2),
the 'make debug' command now halts the chip and waits for a gdb
connection before proceeding execution.