Add wiring_limits.h
This file is analogous to the standard C header <limits.h>.  It
defines various constants within the Wiring language that we consider
to be implementation-defined.  The current version is a work in
progress; more constants may be added as time goes on.

The member constants defined in wiring_limits.h are currently limited
to the core Wiring language.  They do not include
implementation-defined values which pertain to the standard Wiring
libraries.