Make it easier to use the libmaple build system externally.
It's been tricky for users to start their own projects while still
using our build system.  The current recommended practice to get this
done involves modifying the top-level Makefile to add their module
into LIBMAPLE_MODULES, and include their own rules.mk.  Editing the
repository Makefile is stupid and shouldn't be a necessary step.

Instead of that, allow the environment to provide the initial value
for LIBMAPLE_MODULES.  This allows users to specify it on the command
line.

Also add WIRISH_PATH to the initial set of "useful" paths.