Doxygen: add the Evil Mangler.
Whenever Doxygen is running on a series header, make it run an awk
script, the Evil Mangler, that pretends the file is enclosed in an
appropriate namespace declaration for the series target.

Doxygen chokes if two structs have the same name. This is a problem
for the series headers, which commonly have data structures with the
same name.  However, if those structs are in different namespaces,
Doxygen has no problems.  We obviously can't use namespaces in C
headers, so use FILTER_PATTERNS to trick Doxygen into thinking they're
there.

Ugly, but I can't think of a better way to handle this.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>