Replace shiftOut(), also fixing a possible bug.
The current shiftOut() is borrowed from Arduino, and is in an LGPL
file. Replace that file with a new MIT-licensed version containing a
new implementation.

The new version brings the clock line LOW before starting, to make
sure that the first pulse is detected if the clock line was previously
HIGH.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>