Prep usb_descriptors.h for removal.
Just change names and things to fit the libmaple style. Its contents
will move to the top-level <libmaple/usb.h> header next.

This API is still unstable, but we need to expose pieces of it in
order to get the Maple-specific pieces of usb_cdcacm.c into Wirish.

Adjust uses of the API appropriately.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>