conf.py: allow doxygen XML output in repository root
It's a pain to have to re-make the XML output for a given release, and
there's a chance it'll change between different versions of
Doxygen. So look for Doxygen XML output in doxygen/xml/ relative to
the repository root.

This allows maintenance branches to check in the XML for their version
of libmaple.

Signed-off-by: Marti Bolivar <mbolivar@leaflabs.com>