stm32f4: 32bit-float-friendly GCC flags
This commit makes 'float' (32-bit) the default for floating point
constants in the source code, instead of the defaults 'double' (64-bit).

Floating point performance is very sensitive to the float vs. double
distinction on 32-bit processors, especially Cortex-M4s with a VFP unit
for floats but not doubles.

In the future it may be wise to make this change for non-hard-fp chips
(aka, all the other STM32s) as well, for behavioral consistancy. The
warning flag ensures that the frequently used parts of the code base
will be less ambiguous, but doesn't cover series-specific code.